Added a notification limit
The amount of notifications per sun event is now limited to 20 each.
When the limit is reached the user is presented with an alert when
trying to add more notifications.
Also changed the wording for the Duplicate Notification alert.
When trying again from the Duplicate Notification alert, the user is
presented with a picker with the initial selection of where they just
were, which is more user friendly.