Fixed the Norway bug!
Now before sun events are added to either upcomingSunrises or
upcomingSunsets, it is checked to make sure it is a valid sun event.
The isValidSunEventForNumDaysFromNow method was added to the SunEvent
class to check for this. The method is called for each sun event before
being added to to the respective NSArray in myDefaults.